requirements.txt
Django
djangorestframework
Django-Verify-Email==3.0.3
django-countries
python-dotenv
intuit-oauth
python-quickbooks
OdooRPC
django-crontab
httpx
openpyxl
pycryptodome
django-allauth
django-tailwind
django-tailwind[reload]
django-qrcode
gunicorn
psycopg2-binary
django-crontab
python-decouple